fix prodigy, as_tensor in fp64

+++ adv_optm/optim/AdaMuon_adv.py
@@ -353,7 +353,7 @@ class AdaMuon_adv(torch.optim.Optimizer):
             step_size = group['lr'] / bias_correction1
 
             if is_compiled:
-                step_size = torch.as_tensor(step_size)
+                step_size = torch.as_tensor(step_size, dtype=torch.float64)
                 adam_step_param = self._compiled_adam_step_parameter
             else:
                 adam_step_param = Muon_AuxAdam._adam_step_parameter
@@ -364,7 +364,7 @@ class AdaMuon_adv(torch.optim.Optimizer):
 
         else: # Muon path
             if is_compiled:
-                lr = torch.as_tensor(group['lr'])
+                lr = torch.as_tensor(group['lr'], dtype=torch.float64)
                 muon_step_param = self._compiled_muon_step_parameter
             else:
                 lr = group['lr']
